# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fernandovalenzuela_api-home-pay_internal_models.Income:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_recurring:
        type: boolean
      period:
        $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
      period_id:
        type: integer
      received_at:
        type: string
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_models.Period:
    properties:
      id:
        type: integer
      month_number:
        type: integer
      year_number:
        type: integer
    type: object
...
  github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API REST para gestión de gastos e ingresos del hogar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: api-home-pay API
  version: "1.0"
paths:
...
  /api/incomes:
    get:
      consumes:
      - application/json
      description: Get all incomes for the authenticated user, optionally filtered
        by period
      parameters:
      - description: Filter by period ID
        in: query
        name: period_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Income'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all incomes
      tags:
      - incomes
    post:
      consumes:
      - application/json
      description: Create a new income
      parameters:
      - description: Income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Income'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create income
      tags:
      - incomes
  /api/incomes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an income by ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete income
      tags:
      - incomes
    get:
      consumes:
      - application/json
      description: Get a specific income by ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get income by ID
      tags:
      - incomes
    put:
      consumes:
      - application/json
      description: Update an existing income
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: integer
      - description: Income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Income'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update income
      tags:
      - incomes
  /api/periods:
    get:
      consumes:
      - application/json
      description: Get all periods for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all periods
      tags:
      - periods
    post:
      consumes:
      - application/json
      description: Create a new period (month/year)
      parameters:
      - description: Period data
        in: body
        name: period
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create period
      tags:
      - periods
  /api/periods/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a period by ID
      parameters:
      - description: Period ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete period
      tags:
      - periods
    get:
      consumes:
      - application/json
      description: Get a specific period by ID
      parameters:
      - description: Period ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get period by ID
      tags:
      - periods
    put:
      consumes:
      - application/json
      description: Update an existing period
      parameters:
      - description: Period ID
        in: path
        name: id
        required: true
        type: integer
      - description: Period data
        in: body
        name: period
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update period
      tags:
      - periods
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"